// src/app/interfaces/paciente.ts
export interface Paciente {
  nombre: string;
  apellido: string;
  rut: string;
  nombre_social?: string;
  nacimiento?: any;
  telefono?: string;
  correo_electronico?: string;
  sexo?: string;
  genero?: string;
  nacionalidad?: string;
  region?: string;
  comuna?: string;
  direccion?: string;
  pueblo_originario?: string;
  prevision?: string;
  cesfam?: string;
  fecha_nacimiento?: string;
}